make it as `app.mount()`
